refactor: extract duplicated submodule initialization logic into helper function
- Moved repeated code for checking and initializing submodules into a new helper function
- Added initialize_submodule_if_needed() to module_registry module to reduce code duplication
- This addresses PR review feedback about maintainability
